pyproject.toml
[project]
name = "innovate2025-demo-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lucide-fasthtml>=0.0.9",
    "python-fasthtml>=0.9.1",
    "shad4fast>=1.3.1",
]
